Add function prologue and epilogue in debug mode
If we get a trap in the rfxcodec_encode_dwt_shift_amd64_sseXx routine,
the back trace will be garbage because the debugger does not know where
on the stack the frame of the function begins.

This patch adds a standard prologue and epilogue to the function, so the
debugger will find the start of the frame from the base pointer. It is
only added if the library is compiled with debug information enabled.